requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
umap-learn
hdbscan
matplotlib
scipy
pyyaml
